Split loader in CJS and ESM versions
